fixed 32bits error
git-svn-id: https://bugengine.googlecode.com/svn/trunk@221 f3700e90-1ba7-11de-b56b-c9554f25afe1

+++ src/engine/core/src/string/istring.cc
@@ -266,7 +266,7 @@ static inline const char* findnext(const char* str, const char *sep, size_t len)
     return result;
 }
 
-static void parse(const char *str, const char *end, const char *sep, istring* buffer, size_t& size)
+static void parse(const char *str, const char *end, const char *sep, istring* buffer, u32& size)
 {
     size_t numsep = strlen(sep);
